used update, and delete

+++ albums_migration.sql
@@ -5,7 +5,7 @@ CREATE TABLE albums(
 	artist VARCHAR(225) NOT NULL,
 	name VARCHAR(255) NOT NULL,
 	release_date YEAR(4) NOT NULL,
-	sales FLOAT(3, 1) NULL, 
+	sales FLOAT(13, 1) NULL, 
 	genre VARCHAR(255) NOT NULL,
 	PRIMARY KEY(id)
 );
